move diagram function to new module
