+++ examples/examples/custom-scheduler.rs
@@ -38,22 +38,23 @@ async fn main() -> ballista_core::error::Result<()> {
         .is_test(true)
         .try_init();
 
-    let mut config: SchedulerConfig = SchedulerConfig::default();
+    let config: SchedulerConfig = SchedulerConfig {
+        // overriding default runtime producer with custom producer
+        // which knows how to create S3 connections
+        override_config_producer: Some(Arc::new(custom_session_config_with_s3_options)),
+        // overriding default session builder, which has custom session configuration
+        // runtime environment and session state.
+        override_session_builder: Some(Arc::new(|session_config: SessionConfig| {
+            custom_session_state_with_s3_support(session_config)
+        })),
+        ..Default::default()
+    };
 
     let addr = format!("{}:{}", config.bind_host, config.bind_port);
     let addr = addr
         .parse()
         .map_err(|e: AddrParseError| BallistaError::Configuration(e.to_string()))?;
 
-    // overriding default runtime producer with custom producer
-    // which knows how to create S3 connections
-    config.override_config_producer =
-        Some(Arc::new(custom_session_config_with_s3_options));
-    // overriding default session builder, which has custom session configuration
-    // runtime environment and session state.
-    config.override_session_builder = Some(Arc::new(|session_config: SessionConfig| {
-        custom_session_state_with_s3_support(session_config)
-    }));
     let cluster = BallistaCluster::new_from_config(&config).await?;
     start_server(cluster, addr, Arc::new(config)).await?;
